Move PDF/log RHs to common.py
